[REFACTOR] #25 - Refactor HomeViewControllerTest to apply Rest Docs
